move macros around
